ControllerをAPIとHtmlで分けるように修正、APIレスポンスでHtmlが返っていた問題を修正
