+++ app/controllers/Todo.scala
@@ -209,105 +209,4 @@ class TodoController @Inject()(
       }
     )
   }
-  def api_getAll() = Action async { implicit req =>
-    val todoRepo = TodoRepository.list()
-    val categoryRepo = CategoryRepository.list()
-    for {
-      todoEmbed <- todoRepo
-      categoryEmbed <- categoryRepo
-    } yield{
-      val json = todoEmbed.map(TodoCategoryJson.write(_,categoryEmbed))
-      Ok(Json.toJson(json))
-    } 
-  }
-  def api_add() = Action async { implicit req =>
-    TodoCategoryJson.form.bindFromRequest.fold(
-      errorform => {
-        val json = ErrorJson.write(errorform.toString)
-        Future.successful(BadRequest(Json.toJson(json)))
-      },
-      successform => {
-        
-        val todo = Todo.apply(
-          Category.Id(successform.categoryId),
-          successform.title,
-          successform.body,
-          Todo.Status.apply(successform.state.toShort)
-        )
-        val todoAddRepo = TodoRepository.add(todo)
-        for {
-          _ <- todoAddRepo
-        } yield {
-          val json = MessageJson.write("登録しました")
-          Ok(Json.toJson(json))
-        }
-      }
-    )
-  }
-  def api_get(id:Long) = Action async { implicit req =>
-    val todoRepo = TodoRepository.get(Todo.Id(id))
-    val categoryRepo = CategoryRepository.list()
-    for {
-      optionTodo <- todoRepo
-      categoryEmbed <- categoryRepo
-    } yield {
-      optionTodo match {
-        case None => {
-          val json = ErrorJson.write("Todo=" + id + " は存在しません。")
-          NotFound(Json.toJson(json))
-        };
-        case Some(todoEmbed) => {
-          val json = TodoCategoryJson.write(todoEmbed,categoryEmbed)
-          Ok(Json.toJson(json))
-        }
-      }
-    }
-  }
-  def api_update(id:Long) = Action async { implicit req =>
-    TodoCategoryJson.form.bindFromRequest.fold(
-      errorform => {
-        val vv = ViewValueError(
-            message = errorform.toString
-          )
-        Future.successful(BadRequest(views.html.error(vv)))
-      },
-      successform => {
-        val todoEmbededId = new Todo(
-          id = Some(Todo.Id(id)),
-          categoryId = Category.Id(successform.categoryId),
-          title = successform.title,
-          body = successform.body,
-          state = Todo.Status.apply(successform.state.toShort),
-          ).toEmbeddedId //EmbededId型に変換
-        val todoUpdateRepo = TodoRepository.update(todoEmbededId)
-        for {
-          todo <- todoUpdateRepo
-        } yield {
-          todo match {
-            case None => {
-              val json = ErrorJson.write("Todo=" + id + " は存在しません。")
-              NotFound(Json.toJson(json))
-            }
-            case Some(_) => {
-              val json = MessageJson.write("Todo=" + id + " を更新しました")
-              Ok(Json.toJson(json))
-            }
-          }
-        }
-      }
-    )
-  }
-  def api_delete(id: Long) = Action async { implicit req =>
-      val todoId = Todo.Id(id)
-      val todoRepo = TodoRepository.remove(todoId)
-      for {
-        todoDelete <- todoRepo
-      } yield {
-        todoDelete match {
-          case _ =>
-            val json = MessageJson.write("Todo=" + id + " を削除しました")
-            Ok(Json.toJson(json))
-        }
-      }
-  }
 }
